Initial file restructuring
