+++ themes/5ePHB/settings.json
@@ -1,0 +1,4 @@
+{
+  "renderer" : "V3",
+  "baseTheme" : false
+}
