+++ themes/5ePHB/snippets/classtable.gen.js
