+++ themes/5ePHB/snippets/fullclass.gen.js
