+++ themes/5ePHB/snippets/magic.gen.js
