+++ themes/5ePHB/snippets/monsterblock.gen.js
