+++ themes/5ePHBLegacy/settings.json
@@ -1,0 +1,4 @@
+{
+  "renderer" : "legacy",
+  "baseTheme" : false
+}
