+++ themes/5ePHBLegacy/snippets/classfeature.gen.js
