+++ themes/5ePHBLegacy/snippets/fullclass.gen.js
