+++ themes/5ePHBLegacy/snippets/magic.gen.js
