+++ themes/5ePHBLegacy/snippets/monsterblock.gen.js
